added rank, finished error handling

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: syukna <syukna@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 17:07:59 by syukna            #+#    #+#             */
-/*   Updated: 2025/02/06 15:57:49 by syukna           ###   ########.fr       */
+/*   Updated: 2025/02/25 12:55:33 by syukna           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ int		ft_tolower(int c);
 char	*ft_strchr(const char *str, int c);
 char	*ft_strrchr(const char *str, int c);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
-int		ft_atoi(const char *str);
+long int	ft_atoi(const char *str);
 void	*ft_memset(void *str, int c, size_t n);
 void	ft_bzero(void *s, size_t n);
 size_t	ft_strlcpy(char *dest, const char *str, size_t n);
